feat グループ概要ページに概要を移植
